Use a glossary for DissociatedIPC.rst
Without this change, the usage of,

:term:`tag <Tag>

later in the document wasn't linking to anything since Tag isn't a term in any glossary Sphinx knows about. You also get this warning when you build the docs,

/Users/bryce/src/apache/arrow/docs/source/format/DissociatedIPC.rst:278: WARNING: term not in glossary: 'Tag'

I didn't think we wanted to define Tag in the site-wide glossary so I use a page-specific glossary here. It renders the same as before but now can be linked to properly